Fix "unreachable intro" bug
This was a fun one. After a minute of play, you couldn't navigate
back into the intro after the album loops.

This was because eventStarted was being stored as uint16_t, which has
a max value of 65535... or 65 seconds.

Upping to 32bit like millis() fixes the bug by extending the window
to ~50 days.

+++ arduino/songs_for_desmond/common.h
@@ -28,7 +28,7 @@ struct State {
   Stage stage;
 
   int8_t trackIndex;
-  uint16_t eventStarted;
+  uint32_t eventStarted;
 
   Volume volume;
